requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
pandas>=1.3.0
Pillow>=8.3.1
opencv-python>=4.5.3
matplotlib>=3.4.3
seaborn>=0.11.2
scikit-learn>=0.24.2
scikit-image>=0.18.2
tensorflow>=2.7.0
efficientnet>=1.0.0
streamlit>=1.0.0
plotly>=5.1.0
shap>=0.41.0
lime>=0.2.0
umap-learn>=0.5.1
tqdm>=4.62.2
